feat:实现环境变量和aux，但busybox还是不能跑，还需要tls相关的不知道什么东西,也是动态链接用的
Former-commit-id: e3fd829ff313e048d7c8fbf3a7da4411a133561e

+++ os/src/syscall/mod.rs
@@ -116,7 +116,7 @@ pub fn syscall(syscall_id: usize, args: [isize; 6]) -> isize {
         SYSCALL_EXECVE => sys_execve(
             args[0] as *const u8,
             args[1] as *const usize,
-            args[2] as *const u8,
+            args[2] as *const usize,
         ),
         SYSCALL_MMAP => sys_mmap(
             args[0] as usize,
